pyproject.toml
[tool.poetry]
name = "meduzzen-be"
version = "0.1.0"
description = ""
authors = ["Svitlana Tsytsyk <TSYTSYK.svetlana@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.115.2"}
uvicorn = "^0.32.0"
aioredis = "^2.0.1"
sqlalchemy = "^2.0.36"
pydantic = "^2.9.2"
asyncpg = "^0.30.0"
redis = "^5.1.1"
alembic = "^1.13.3"
black = "^24.10.0"
python-json-logger = "^2.0.7"
bcrypt = "^4.2.0"
asyncio = "^3.4.3"
passlib = "^1.7.4"
fastapi-jwt = "^0.3.0"
python-jose = "^3.3"
pydantic-settings = "^2.6.0"
requests = "^2.32.3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pre-commit = "^4.0.1"
anyio = "^4.6.2.post1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
